Prevent script from dying if communication with the tuner fails
Periodically the script will fail while scanning, because the communication
with the tuner fails. I'm not sure if this is due to problems with the tuner,
the network, libhdhomerun, or something else. The stack trace looks like this:

/usr/local/rvm/gems/ruby-1.9.3-p448/gems/ffi-hdhomerun-0.6/lib/ffi-hdhomerun.rb:546:in `get': communication error sending request to hdhomerun device (RuntimeError)
        from /usr/local/rvm/gems/ruby-1.9.3-p448/gems/ffi-hdhomerun-0.6/lib/ffi-hdhomerun.rb:570:in `get_tuner'
        from /usr/local/rvm/gems/ruby-1.9.3-p448/gems/ffi-hdhomerun-0.6/lib/ffi-hdhomerun.rb:526:in `locked?'
        from /usr/local/rvm/gems/ruby-1.9.3-p448/gems/ffi-hdhomerun-0.6/lib/ffi-hdhomerun.rb:307:in `scan'
        from ./scan.rb:35:in `scan'
        from ./scan.rb:294:in `block (3 levels) in <main>'

I've wrapped the entire can block in a rescue block so we can dump out something
if we get an error, but continue along in any case.